+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- Add method to originate channels.
+- Add methods controling a `LivePlayback`.
+
 ## 0.1.0
 
 - Initial version.